fix(username): fix no username after login
